Add command giver test for early blasting of towers (#44)
- Add check for bot transform at spawn position

+++ src/constants/include/constants/actor.h
@@ -21,22 +21,22 @@ const size_t MAX_NUM_BOTS = 500;
 const size_t NUM_BOTS_START = 20;
 
 // Blast impact radius around the bot
-const size_t BOT_BLAST_IMPACT_RADIUS = 3;
+const uint64_t BOT_BLAST_IMPACT_RADIUS = 3;
 
 // Speed of a bot in manhattan distance per turn
-const int64_t BOT_SPEED = 2;
+const uint64_t BOT_SPEED = 2;
 
 // Frequency at which bots spawn at home base per turn
-const int64_t BOT_SPAWN_FREQUENCY = 1;
+const uint64_t BOT_SPAWN_FREQUENCY = 1;
 
 // Maximum number of towers per player
 const size_t MAX_NUM_TOWERS = 200;
 
 // Blast impact radius around the tower
-const size_t TOWER_BLAST_IMPACT_RADIUS = 3;
+const uint64_t TOWER_BLAST_IMPACT_RADIUS = 3;
 
 // Minimum number of turns tower must be alive to be able to blast
-const size_t TOWER_MIN_BLAST_AGE = 5;
+const uint64_t TOWER_MIN_BLAST_AGE = 5;
 
 } // namespace Constants::Actor
